source/requirements.txt
transformers==4.41.2
accelerate
datasets
torchvision==0.18.0
wandb
evaluate
nltk
torch>=2.1.0
pillow
scikit-learn
sentencepiece
